refactor #112: move badges outside table with controls component
