+++ components/badges/__tests__/index.tsx
