+++ components/badges/index.tsx
